Додав файл main.rb

+++ main.rb
@@ -1,9 +1,9 @@
 puts "Введи перше число:"
-a = gets.to_f
+a = gets.to_i
 puts "Введи оператор (+, -, *, /):"
 b = gets.chomp
 puts "Введи друге число:"
-c = gets.to_f
+c = gets.to_i
 
 result = a.send(b, c) rescue "Помилка"
